@@ -1,24 +1,24 @@
 module github.com/Huawei/eSDK_K8S_Plugin/v4
 
-go 1.22
+go 1.24.1
 
 require (
 	github.com/agiledragon/gomonkey/v2 v2.9.0
 	github.com/container-storage-interface/spec v1.8.0
-	github.com/golang/protobuf v1.5.3
+	github.com/golang/protobuf v1.5.4
 	github.com/kubernetes-csi/csi-lib-utils v0.11.0
 	github.com/prashantv/gostub v1.1.0
-	github.com/sirupsen/logrus v1.8.2
+	github.com/sirupsen/logrus v1.9.3
 	github.com/spf13/cobra v1.8.0
-	github.com/stretchr/testify v1.9.0
+	github.com/stretchr/testify v1.10.0
 	go.uber.org/mock v0.5.0
 	golang.org/x/sys v0.24.0
 	google.golang.org/grpc v1.57.2
-	google.golang.org/protobuf v1.32.0
+	google.golang.org/protobuf v1.34.2
 	gopkg.in/yaml.v3 v3.0.1
-	k8s.io/api v0.29.2
-	k8s.io/apimachinery v0.29.2
-	k8s.io/client-go v0.29.2
+	k8s.io/api v0.31.1
+	k8s.io/apimachinery v0.31.1
+	k8s.io/client-go v0.31.1
 	k8s.io/code-generator v0.27.6
 )
 
