@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) Huawei Technologies Co., Ltd. 2024-2024. All rights reserved.
+ *  Copyright (c) Huawei Technologies Co., Ltd. 2024-2025. All rights reserved.
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -122,4 +122,13 @@ func (p *OceandiskPlugin) Logout(ctx context.Context) {
 	if p.cli != nil {
 		p.cli.Logout(ctx)
 	}
+}
+
+// ReLogin will refresh the user session of storage
+func (p *OceandiskPlugin) ReLogin(ctx context.Context) error {
+	if p.cli == nil {
+		return nil
+	}
+
+	return p.cli.ReLogin(ctx)
 }
