@@ -14,11 +14,12 @@
  *  limitations under the License.
  */
 
-// Package delete_volume includes the integration tests of creating volume
+// Package create_volume includes the integration tests of creating volume
 package create_volume
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
 	"testing"
 
@@ -34,6 +35,7 @@ import (
 	"github.com/Huawei/eSDK_K8S_Plugin/v4/csi/backend/model"
 	"github.com/Huawei/eSDK_K8S_Plugin/v4/csi/backend/plugin"
 	"github.com/Huawei/eSDK_K8S_Plugin/v4/pkg/constants"
+	pkgUtils "github.com/Huawei/eSDK_K8S_Plugin/v4/pkg/utils"
 	"github.com/Huawei/eSDK_K8S_Plugin/v4/storage/oceanstorage/oceanstor/client"
 	"github.com/Huawei/eSDK_K8S_Plugin/v4/test/mocks/mock_client"
 	"github.com/Huawei/eSDK_K8S_Plugin/v4/test/utils"
@@ -45,14 +47,14 @@ func TestCreateVolume_OceanstorDTree_FullFeaturesSuccess(t *testing.T) {
 	ctx := context.Background()
 	mockCtrl := gomock.NewController(t)
 	cli := mock_client.NewMockOceanstorClientInterface(mockCtrl)
-	cache.BackendCacheProvider.Store(ctx, data.BackendName, data.backend(cli))
+	cache.BackendCacheProvider.Store(ctx, data.BackendName, data.backend(cli, constants.OceanStorDoradoV6))
 	defer cache.BackendCacheProvider.Delete(ctx, data.BackendName)
 
 	// mock
 	p := gomonkey.ApplyMethodReturn(app.GetGlobalConfig().K8sUtils, "GetVolumeConfiguration", map[string]string{}, nil)
 	defer p.Reset()
 	cli.EXPECT().GetFileSystemByName(ctx, data.ExpectedParentName).Return(map[string]any{"ID": data.FakeFsID}, nil)
-	cli.EXPECT().CreateDTree(ctx, data.expectedCreateDTreeParams()).Return(map[string]any{"ID": data.FakeDTreeID}, nil)
+	cli.EXPECT().CreateDTree(ctx, data.expectedCreateDTreeParams(t)).Return(map[string]any{"ID": data.FakeDTreeID}, nil)
 	cli.EXPECT().GetNfsShareByPath(ctx, data.expectedSharePath(), data.FakeVStoreID).Return(nil, nil)
 	cli.EXPECT().CreateNfsShare(ctx, data.expectedCreateNfsShareParams()).
 		Return(map[string]any{"ID": data.FakeShareID}, nil)
@@ -69,13 +71,114 @@ func TestCreateVolume_OceanstorDTree_FullFeaturesSuccess(t *testing.T) {
 	require.Equal(t, data.response(), resp)
 }
 
+func TestCreateVolume_OceanstorDTree_SuccessWithScVolumeName(t *testing.T) {
+	// arrange
+	data := fakeOceanstorDtreeDataWithSuccess()
+	data.ScVolumeName = "prefix-{{ .PVCNamespace }}-{{ .PVCName }}"
+	data.ExpectedDTreeName = "prefix-pvcTestNamespace-pvcTestName-pvTestName"
+	ctx := context.Background()
+	mockCtrl := gomock.NewController(t)
+	cli := mock_client.NewMockOceanstorClientInterface(mockCtrl)
+	cache.BackendCacheProvider.Store(ctx, data.BackendName, data.backend(cli, constants.OceanStorDoradoV6))
+	defer cache.BackendCacheProvider.Delete(ctx, data.BackendName)
+	createVolumeReq := data.request()
+	createVolumeReq.Parameters[constants.PVCNamespaceKey] = "pvcTestNamespace"
+	createVolumeReq.Parameters[constants.PVCNameKey] = "pvcTestName"
+	createVolumeReq.Parameters[constants.PVNameKey] = "pvc-pvTestName"
+	originPrefix := app.GetGlobalConfig().VolumeNamePrefix
+	defer func() { app.GetGlobalConfig().VolumeNamePrefix = originPrefix }()
+	app.GetGlobalConfig().VolumeNamePrefix = "pvc-"
+
+	// mock
+	p := gomonkey.ApplyMethodReturn(app.GetGlobalConfig().K8sUtils, "GetVolumeConfiguration", map[string]string{}, nil)
+	defer p.Reset()
+	cli.EXPECT().GetFileSystemByName(ctx, data.ExpectedParentName).Return(map[string]any{"ID": data.FakeFsID}, nil)
+	cli.EXPECT().CreateDTree(ctx, data.expectedCreateDTreeParams(t)).Return(map[string]any{"ID": data.FakeDTreeID}, nil)
+	cli.EXPECT().GetNfsShareByPath(ctx, data.expectedSharePath(), data.FakeVStoreID).Return(nil, nil)
+	cli.EXPECT().CreateNfsShare(ctx, data.expectedCreateNfsShareParams()).
+		Return(map[string]any{"ID": data.FakeShareID}, nil)
+	cli.EXPECT().GetNfsShareAccessCount(ctx, data.FakeShareID, data.FakeVStoreID).Return(int64(0), nil)
+	cli.EXPECT().AllowNfsShareAccess(ctx, data.expectedAllowNfsShareRequest()).Return(nil)
+	cli.EXPECT().CreateQuota(ctx, data.expectedCreateQuotaParam()).Return(nil, nil)
+	cli.EXPECT().Logout(ctx)
+
+	// action
+	resp, err := csiServer.CreateVolume(ctx, createVolumeReq)
+
+	// assert
+	require.NoError(t, err)
+	require.Equal(t, data.response(), resp)
+}
+
+func TestCreateVolume_OceanstorDTree_FailedWithInvalidScVolumeName(t *testing.T) {
+	// arrange
+	data := fakeOceanstorDtreeDataWithSuccess()
+	data.ScVolumeName = "prefix-{{ .PVCNamespace }}"
+	ctx := context.Background()
+	mockCtrl := gomock.NewController(t)
+	cli := mock_client.NewMockOceanstorClientInterface(mockCtrl)
+	cache.BackendCacheProvider.Store(ctx, data.BackendName, data.backend(cli, constants.OceanStorDoradoV6))
+	defer cache.BackendCacheProvider.Delete(ctx, data.BackendName)
+	createVolumeReq := data.request()
+
+	// mock
+	p := gomonkey.ApplyMethodReturn(app.GetGlobalConfig().K8sUtils, "GetVolumeConfiguration", map[string]string{}, nil)
+	defer p.Reset()
+	cli.EXPECT().Logout(ctx)
+
+	// action
+	_, err := csiServer.CreateVolume(ctx, createVolumeReq)
+
+	// assert
+	require.ErrorContains(t, err,
+		"{{.PVCNamespace}} or {{.PVCName}} must be configured in the volumeName parameter at the same time")
+}
+
+func TestCreateVolume_OceanstorDTreeV5_SuccessWithScVolumeName(t *testing.T) {
+	// arrange
+	data := fakeOceanstorDtreeDataWithSuccess()
+	data.ScVolumeName = "prefix-{{ .PVCNamespace }}-{{ .PVCName }}"
+	ctx := context.Background()
+	mockCtrl := gomock.NewController(t)
+	cli := mock_client.NewMockOceanstorClientInterface(mockCtrl)
+	cache.BackendCacheProvider.Store(ctx, data.BackendName, data.backend(cli, constants.OceanStorV5))
+	defer cache.BackendCacheProvider.Delete(ctx, data.BackendName)
+	createVolumeReq := data.request()
+	createVolumeReq.Parameters[constants.PVCNamespaceKey] = "pvcTestNamespace"
+	createVolumeReq.Parameters[constants.PVCNameKey] = "pvcTestName"
+	createVolumeReq.Parameters[constants.PVNameKey] = "pvc-pvTestName"
+	originPrefix := app.GetGlobalConfig().VolumeNamePrefix
+	defer func() { app.GetGlobalConfig().VolumeNamePrefix = originPrefix }()
+	app.GetGlobalConfig().VolumeNamePrefix = "pvc-"
+
+	// mock
+	p := gomonkey.ApplyMethodReturn(app.GetGlobalConfig().K8sUtils, "GetVolumeConfiguration", map[string]string{}, nil)
+	defer p.Reset()
+	cli.EXPECT().GetFileSystemByName(ctx, data.ExpectedParentName).Return(map[string]any{"ID": data.FakeFsID}, nil)
+	cli.EXPECT().CreateDTree(ctx, data.expectedCreateDTreeParams(t)).Return(map[string]any{"ID": data.FakeDTreeID}, nil)
+	cli.EXPECT().GetNfsShareByPath(ctx, data.expectedSharePath(), data.FakeVStoreID).Return(nil, nil)
+	cli.EXPECT().CreateNfsShare(ctx, data.expectedCreateNfsShareParams()).
+		Return(map[string]any{"ID": data.FakeShareID}, nil)
+	cli.EXPECT().GetNfsShareAccessCount(ctx, data.FakeShareID, data.FakeVStoreID).Return(int64(0), nil)
+	cli.EXPECT().AllowNfsShareAccess(ctx, data.expectedAllowNfsShareRequest()).Return(nil)
+	cli.EXPECT().CreateQuota(ctx, data.expectedCreateQuotaParam()).Return(nil, nil)
+	cli.EXPECT().Logout(ctx)
+
+	// action
+	resp, err := csiServer.CreateVolume(ctx, createVolumeReq)
+
+	// assert
+	require.NoError(t, err)
+	require.Equal(t, data.response(), resp)
+}
+
 func TestCreateVolume_OceanstorDTree_FailedWithInvalidCapacity(t *testing.T) {
 	// arrange
 	data := fakeOceanstorDTreeDataWithInvalidCapacity()
 	ctx := context.Background()
 	mockCtrl := gomock.NewController(t)
 	cli := mock_client.NewMockOceanstorClientInterface(mockCtrl)
-	cache.BackendCacheProvider.Store(ctx, data.BackendName, data.backend(cli))
+	cache.BackendCacheProvider.Store(ctx, data.BackendName, data.backend(cli, constants.OceanStorDoradoV6))
 	defer cache.BackendCacheProvider.Delete(ctx, data.BackendName)
 
 	// mock
@@ -97,7 +200,7 @@ func TestCreateVolume_OceanstorDTree_FailedWithEmptyParentName(t *testing.T) {
 	ctx := context.Background()
 	mockCtrl := gomock.NewController(t)
 	cli := mock_client.NewMockOceanstorClientInterface(mockCtrl)
-	cache.BackendCacheProvider.Store(ctx, data.BackendName, data.backend(cli))
+	cache.BackendCacheProvider.Store(ctx, data.BackendName, data.backend(cli, constants.OceanStorDoradoV6))
 	defer cache.BackendCacheProvider.Delete(ctx, data.BackendName)
 
 	// mock
@@ -119,7 +222,7 @@ func TestCreateVolume_OceanstorDTree_FailedWithDifferentParentName(t *testing.T)
 	ctx := context.Background()
 	mockCtrl := gomock.NewController(t)
 	cli := mock_client.NewMockOceanstorClientInterface(mockCtrl)
-	cache.BackendCacheProvider.Store(ctx, data.BackendName, data.backend(cli))
+	cache.BackendCacheProvider.Store(ctx, data.BackendName, data.backend(cli, constants.OceanStorDoradoV6))
 	defer cache.BackendCacheProvider.Delete(ctx, data.BackendName)
 
 	// mock
@@ -135,6 +238,29 @@ func TestCreateVolume_OceanstorDTree_FailedWithDifferentParentName(t *testing.T)
 		fmt.Sprintf("parentname %q in StorageClass is not equal to %q in backend", data.ScParentName,
 			data.BackendParentName))
 	require.Nil(t, resp)
+}
+
+func TestCreateVolume_OceanstorDTree_AdvancedOptionsUnmarshalFailed(t *testing.T) {
+	// arrange
+	data := fakeOceanstorDtreeDataWithSuccess()
+	data.AdvancedOptions = `{"CAPACITYTHRESHOLD": 90`
+	ctx := context.Background()
+	mockCtrl := gomock.NewController(t)
+	cli := mock_client.NewMockOceanstorClientInterface(mockCtrl)
+	cache.BackendCacheProvider.Store(ctx, data.BackendName, data.backend(cli, constants.OceanStorDoradoV6))
+	defer cache.BackendCacheProvider.Delete(ctx, data.BackendName)
+
+	// mock
+	p := gomonkey.ApplyMethodReturn(app.GetGlobalConfig().K8sUtils, "GetVolumeConfiguration", map[string]string{}, nil)
+	defer p.Reset()
+	cli.EXPECT().GetFileSystemByName(ctx, data.ExpectedParentName).Return(map[string]any{"ID": data.FakeFsID}, nil)
+	cli.EXPECT().Logout(ctx)
+
+	// action
+	_, err := csiServer.CreateVolume(ctx, data.request())
+
+	// assert
+	require.ErrorContains(t, err, "failed to unmarshal advancedOptions")
 }
 
 func fakeOceanstorDTreeDataWithDifferentParentName() *oceanstorDTree {
@@ -192,10 +318,12 @@ func fakeOceanstorDtreeDataWithSuccess() *oceanstorDTree {
 		AccessKrb5:            "read_write",
 		AccessKrb5i:           "read_write",
 		AccessKrb5p:           "read_write",
+		AdvancedOptions:       `{"CAPACITYTHRESHOLD": 90}`,
 
 		BackendName:       "test-dtree-backend",
 		BackendParentName: "test-backend-parent-name",
 
+		ExpectedDTreeName:        "pvc-test-dtree",
 		ExpectedParentName:       "test-backend-parent-name",
 		ExpectedAllSquashParam:   0,
 		ExpectedRootSquashParam:  0,
@@ -225,10 +353,13 @@ type oceanstorDTree struct {
 	AccessKrb5i           string `sc:"accesskrb5i"`
 	AccessKrb5p           string `sc:"accesskrb5p"`
 	ScParentName          string `sc:"parentname"`
+	AdvancedOptions       string `sc:"advancedOptions"`
+	ScVolumeName          string `sc:"volumeName"`
 
 	BackendName       string `sc:"backend"`
 	BackendParentName string
 
+	ExpectedDTreeName        string
 	ExpectedParentName       string
 	ExpectedAllSquashParam   int
 	ExpectedRootSquashParam  int
@@ -242,10 +373,12 @@ type oceanstorDTree struct {
 	FakeShareID  string
 }
 
-func (dtree *oceanstorDTree) backend(cli client.OceanstorClientInterface) model.Backend {
+func (dtree *oceanstorDTree) backend(cli client.OceanstorClientInterface,
+	product constants.OceanstorVersion) model.Backend {
 	p := &plugin.OceanstorDTreePlugin{}
 	p.SetCli(cli)
 	p.SetParentName(dtree.BackendParentName)
+	p.SetProduct(product)
 	return model.Backend{
 		Name:        dtree.BackendName,
 		ContentName: "test-content-name",
@@ -306,10 +439,10 @@ func (dtree *oceanstorDTree) response() *csi.CreateVolumeResponse {
 	return &csi.CreateVolumeResponse{
 		Volume: &csi.Volume{
 			CapacityBytes: dtree.Capacity,
-			VolumeId:      dtree.BackendName + "." + dtree.Name,
+			VolumeId:      dtree.BackendName + "." + dtree.ExpectedDTreeName,
 			VolumeContext: map[string]string{
 				"backend":                          dtree.BackendName,
-				"name":                             dtree.Name,
+				"name":                             dtree.ExpectedDTreeName,
 				"fsPermission":                     dtree.Permission,
 				constants.DTreeParentKey:           dtree.BackendParentName,
 				constants.DisableVerifyCapacityKey: dtree.DisableVerifyCapacity,
@@ -342,7 +475,7 @@ func (dtree *oceanstorDTree) expectedAllowNfsShareRequest() *client.AllowNfsShar
 		AccessVal:   1,
 		Sync:        0,
 		AllSquash:   dtree.ExpectedAllSquashParam,
-		RootSquash:  dtree.ExpectedAllSquashParam,
+		RootSquash:  dtree.ExpectedRootSquashParam,
 		VStoreID:    dtree.FakeVStoreID,
 		AccessKrb5:  dtree.ExpectedAccessKrb5Param,
 		AccessKrb5i: dtree.ExpectedAccessKrb5iParam,
@@ -361,15 +494,22 @@ func (dtree *oceanstorDTree) expectedCreateNfsShareParams() map[string]any {
 }
 
 func (dtree *oceanstorDTree) expectedSharePath() string {
-	return fmt.Sprintf("/%s/%s", dtree.BackendParentName, dtree.Name)
+	return fmt.Sprintf("/%s/%s", dtree.BackendParentName, dtree.ExpectedDTreeName)
 }
 
-func (dtree *oceanstorDTree) expectedCreateDTreeParams() map[string]any {
-	return map[string]any{
+func (dtree *oceanstorDTree) expectedCreateDTreeParams(t *testing.T) map[string]any {
+	params := map[string]any{
 		"unixPermissions": dtree.Permission,
-		"NAME":            dtree.Name,
+		"NAME":            dtree.ExpectedDTreeName,
 		"PARENTNAME":      dtree.BackendParentName,
 		"PARENTTYPE":      client.ParentTypeFS,
 		"securityStyle":   client.SecurityStyleUnix,
 	}
+	if dtree.AdvancedOptions != "" {
+		advancedOptions := make(map[string]any)
+		err := json.Unmarshal([]byte(dtree.AdvancedOptions), &advancedOptions)
+		require.NoError(t, err)
+		params = pkgUtils.CombineMap(params, advancedOptions)
+	}
+	return params
 }
